--- interfaces/index.ts
@@ -1,3 +1,0 @@
-export * from './discord.interface'
-export * from './supabase.interface'
-export * from './xbox.interface'
